+++ compiler-core/src/io.rs
@@ -66,6 +66,30 @@ impl Content {
     }
 }
 
+impl From<Vec<u8>> for Content {
+    fn from(bytes: Vec<u8>) -> Self {
+        Content::Binary(bytes)
+    }
+}
+
+impl From<&[u8]> for Content {
+    fn from(bytes: &[u8]) -> Self {
+        Content::Binary(bytes.to_vec())
+    }
+}
+
+impl From<String> for Content {
+    fn from(text: String) -> Self {
+        Content::Text(text)
+    }
+}
+
+impl From<&str> for Content {
+    fn from(text: &str) -> Self {
+        Content::Text(text.to_string())
+    }
+}
+
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct OutputFile {
     pub content: Content,
@@ -137,6 +161,7 @@ pub trait FileSystemReader {
     fn gleam_metadata_files(&self, dir: &Path) -> Vec<PathBuf>;
     fn read_dir(&self, path: &Path) -> Result<ReadDir>;
     fn read(&self, path: &Path) -> Result<String, Error>;
+    fn read_bytes(&self, path: &Path) -> Result<Vec<u8>, Error>;
     fn reader(&self, path: &Path) -> Result<WrappedReader, Error>;
     fn is_file(&self, path: &Path) -> bool;
     fn is_directory(&self, path: &Path) -> bool;