+++ compiler-core/src/javascript/tests.rs
@@ -120,10 +120,10 @@ macro_rules! assert_js {
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = vec!["my".to_string(), "mod".to_string()];
         let ast = $crate::type_::infer_module(
-            crate::build::Target::JavaScript,
+            $crate::build::Target::JavaScript,
             &ids,
             ast,
-            crate::build::Origin::Src,
+            $crate::build::Origin::Src,
             "thepackage",
             &modules,
             &mut vec![],
@@ -135,8 +135,8 @@ macro_rules! assert_js {
     }};
 
     ($src:expr, $js:expr $(,)?) => {{
-        use crate::{javascript::*, uid::UniqueIdGenerator};
         use std::path::Path;
+        use $crate::{javascript::*, uid::UniqueIdGenerator};
         let mut modules = im::HashMap::new();
         let ids = UniqueIdGenerator::new();
         // DUPE: preludeinsertion
@@ -257,10 +257,10 @@ macro_rules! assert_ts_def {
         let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = vec!["my".to_string(), "mod".to_string()];
         let ast = $crate::type_::infer_module(
-            crate::build::Target::JavaScript,
+            $crate::build::Target::JavaScript,
             &ids,
             ast,
-            crate::build::Origin::Src,
+            $crate::build::Origin::Src,
             "thepackage",
             &modules,
             &mut vec![],
@@ -271,8 +271,8 @@ macro_rules! assert_ts_def {
     }};
 
     ($src:expr, $js:expr $(,)?) => {{
-        use crate::{javascript::*, uid::UniqueIdGenerator};
         use std::path::Path;
+        use $crate::{javascript::*, uid::UniqueIdGenerator};
         let mut modules = im::HashMap::new();
         let ids = UniqueIdGenerator::new();
         // DUPE: preludeinsertion