+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - Scientific notation is now available for float literals.
+- The compiler now supports incremental compilation at the module level. If a
+  module or its dependencies have not been changed unchanged then it will not be
+  recompiled.
 - Fixed a bug where no error would be emitted if a `src` module imported a
   `test` module.
 - Fixed a bug where comments in list prepending expressions could be formatted